+++ crates/litesvm/benches/simple_bench.rs
@@ -2,13 +2,13 @@ use std::path::PathBuf;
 
 use criterion::{criterion_group, criterion_main, Criterion};
 use litesvm::LiteSVM;
-use solana_program::{
-    instruction::{AccountMeta, Instruction},
-    pubkey::Pubkey,
+use {
+    solana_account::Account, solana_keypair::Keypair, solana_message::Message,
+    solana_signer::Signer, solana_transaction::Transaction,
 };
-use solana_sdk::{
-    account::Account, message::Message, signature::Keypair, signer::Signer,
-    transaction::Transaction,
+use {
+    solana_instruction::{account_meta::AccountMeta, Instruction},
+    solana_pubkey::Pubkey,
 };
 
 fn read_counter_program() -> Vec<u8> {
@@ -23,7 +23,7 @@ fn make_tx(
     program_id: Pubkey,
     counter_address: Pubkey,
     payer_pk: &Pubkey,
-    blockhash: solana_program::hash::Hash,
+    blockhash: solana_hash::Hash,
     payer_kp: &Keypair,
     deduper: u8,
 ) -> Transaction {
@@ -72,7 +72,7 @@ fn criterion_benchmark(c: &mut Criterion) {
     });
 }
 
-fn counter_acc(program_id: Pubkey) -> solana_sdk::account::Account {
+fn counter_acc(program_id: Pubkey) -> solana_account::Account {
     Account {
         lamports: 5,
         data: vec![0_u8; std::mem::size_of::<u32>()],