+++ crates/litesvm/src/lib.rs
@@ -23,8 +23,12 @@ In a further break from tradition, it has an ergonomic API with sane defaults an
 
 ```rust
 use litesvm::LiteSVM;
-use solana_program::{message::Message, pubkey::Pubkey, system_instruction::transfer};
-use solana_sdk::{signature::Keypair, signer::Signer, transaction::Transaction};
+use solana_message::Message;
+use solana_pubkey::Pubkey;
+use solana_system_interface::instruction::transfer;
+use solana_keypair::Keypair;
+use solana_signer::Signer;
+use solana_transaction::Transaction;
 
 let from_keypair = Keypair::new();
 let from = from_keypair.pubkey();
@@ -63,12 +67,11 @@ from the Solana Program Library that just does some logging:
 use {
     litesvm::LiteSVM,
     solana_instruction::{account_meta::AccountMeta, Instruction},
+    solana_keypair::Keypair,
     solana_pubkey::{pubkey, Pubkey},
-    solana_sdk::{
-        message::{Message, VersionedMessage},
-        signer::{keypair::Keypair, Signer},
-        transaction::VersionedTransaction,
-    },
+    solana_message::{Message, VersionedMessage},
+    solana_signer::Signer,
+    solana_transaction::versioned::VersionedTransaction,
 };
 
 fn test_logging() {
@@ -114,12 +117,11 @@ use {
     litesvm::LiteSVM,
     solana_clock::Clock,
     solana_instruction::Instruction,
+    solana_keypair::Keypair,
+    solana_message::{Message, VersionedMessage},
     solana_pubkey::Pubkey,
-    solana_sdk::{
-        message::{Message, VersionedMessage},
-        signer::{keypair::Keypair, Signer},
-        transaction::VersionedTransaction,
-    },
+    solana_signer::Signer,
+    solana_transaction::versioned::VersionedTransaction,
 };
 
 fn test_set_clock() {
@@ -260,47 +262,18 @@ use solana_bpf_loader_program::syscalls::create_program_runtime_environment_v1;
 use solana_bpf_loader_program::syscalls::create_program_runtime_environment_v2;
 use solana_compute_budget::compute_budget::ComputeBudget;
 use solana_compute_budget::compute_budget_limits::ComputeBudgetLimits;
+use solana_compute_budget_instruction::instructions_processor::process_compute_budget_instructions;
 use solana_log_collector::LogCollector;
-#[allow(deprecated)]
-use solana_program::sysvar::{fees::Fees, recent_blockhashes::RecentBlockhashes};
 use solana_program_runtime::{
     invoke_context::{BuiltinFunctionWithContext, EnvironmentConfig, InvokeContext},
     loaded_programs::{LoadProgramMetrics, ProgramCacheEntry},
 };
-use solana_runtime_transaction::instructions_processor::process_compute_budget_instructions;
-#[allow(deprecated)]
-use solana_sdk::sysvar::recent_blockhashes::IterItem;
-use solana_sdk::{
-    account::{Account, AccountSharedData, ReadableAccount, WritableAccount},
-    bpf_loader,
-    clock::Clock,
-    epoch_rewards::EpochRewards,
-    epoch_schedule::EpochSchedule,
-    feature_set::{remove_rounding_in_fee_calculation, FeatureSet},
-    fee::FeeStructure,
-    hash::Hash,
-    inner_instruction::InnerInstructionsList,
-    message::{Message, SanitizedMessage, VersionedMessage},
-    native_loader,
-    native_token::LAMPORTS_PER_SOL,
-    nonce::{state::DurableNonce, NONCED_TX_MARKER_IX_INDEX},
-    nonce_account,
-    pubkey::Pubkey,
-    rent::Rent,
-    reserved_account_keys::ReservedAccountKeys,
-    signature::{Keypair, Signature},
-    signer::Signer,
-    slot_hashes::SlotHashes,
-    slot_history::SlotHistory,
-    stake_history::StakeHistory,
-    system_instruction, system_program,
-    sysvar::{last_restart_slot::LastRestartSlot, Sysvar, SysvarId},
-    transaction::{MessageHash, SanitizedTransaction, TransactionError, VersionedTransaction},
-    transaction_context::{ExecutionRecord, IndexOfAccount, TransactionContext},
-};
-use solana_svm::message_processor::MessageProcessor;
 use solana_svm_transaction::svm_message::SVMMessage;
 use solana_system_program::{get_system_account_kind, SystemAccountKind};
+#[allow(deprecated)]
+use solana_sysvar::recent_blockhashes::IterItem;
+#[allow(deprecated)]
+use solana_sysvar::{fees::Fees, recent_blockhashes::RecentBlockhashes};
 use solana_timings::ExecuteTimings;
 use std::{cell::RefCell, path::Path, rc::Rc, sync::Arc};
 use types::SimulatedTransactionInfo;
@@ -308,12 +281,46 @@ use utils::{
     construct_instructions_account,
     inner_instructions::inner_instructions_list_from_instruction_trace,
 };
+use {
+    solana_account::{Account, AccountSharedData, ReadableAccount, WritableAccount},
+    solana_clock::Clock,
+    solana_epoch_rewards::EpochRewards,
+    solana_epoch_schedule::EpochSchedule,
+    solana_feature_set::FeatureSet,
+    solana_fee_structure::FeeStructure,
+    solana_hash::Hash,
+    solana_keypair::Keypair,
+    solana_last_restart_slot::LastRestartSlot,
+    solana_message::{
+        inner_instruction::InnerInstructionsList, Message, SanitizedMessage, VersionedMessage,
+    },
+    solana_native_token::LAMPORTS_PER_SOL,
+    solana_nonce::{state::DurableNonce, NONCED_TX_MARKER_IX_INDEX},
+    solana_pubkey::Pubkey,
+    solana_rent::Rent,
+    solana_reserved_account_keys::ReservedAccountKeys,
+    solana_sdk_ids::{bpf_loader, native_loader, system_program},
+    solana_signature::Signature,
+    solana_signer::Signer,
+    solana_slot_hashes::SlotHashes,
+    solana_slot_history::SlotHistory,
+    solana_stake_interface::stake_history::StakeHistory,
+    solana_sysvar::Sysvar,
+    solana_sysvar_id::SysvarId,
+    solana_transaction::{
+        sanitized::{MessageHash, SanitizedTransaction},
+        versioned::VersionedTransaction,
+    },
+    solana_transaction_context::{ExecutionRecord, IndexOfAccount, TransactionContext},
+    solana_transaction_error::TransactionError,
+};
 
 use crate::{
     accounts_db::AccountsDb,
     builtin::BUILTINS,
     error::LiteSVMError,
     history::TransactionHistory,
+    message_processor::process_message,
     spl::load_spl_programs,
     types::{ExecutionResult, FailedTransactionMetadata, TransactionMetadata, TransactionResult},
     utils::{create_blockhash, rent::RentState},
@@ -326,6 +333,7 @@ mod accounts_db;
 mod builtin;
 mod format_logs;
 mod history;
+mod message_processor;
 mod precompiles;
 mod spl;
 mod utils;
@@ -450,7 +458,7 @@ impl LiteSVM {
                 .replenish(builtint.program_id, Arc::new(loaded_program));
             self.accounts.add_builtin_account(
                 builtint.program_id,
-                native_loader::create_loadable_account_for_test(builtint.name),
+                crate::utils::create_loadable_account_for_test(builtint.name),
             );
 
             if let Some(feature_id) = builtint.feature_id {
@@ -574,7 +582,8 @@ impl LiteSVM {
     where
         T: Sysvar + SysvarId,
     {
-        let account = AccountSharedData::new_data(1, &sysvar, &solana_sdk::sysvar::id()).unwrap();
+        let account =
+            AccountSharedData::new_data(1, &sysvar, &solana_sdk_ids::sysvar::id()).unwrap();
         self.accounts.add_account(T::id(), account).unwrap();
     }
 
@@ -596,7 +605,7 @@ impl LiteSVM {
         let payer = Keypair::from_bytes(&self.airdrop_kp).unwrap();
         let tx = VersionedTransaction::try_new(
             VersionedMessage::Legacy(Message::new_with_blockhash(
-                &[system_instruction::transfer(
+                &[solana_system_interface::instruction::transfer(
                     &payer.pubkey(),
                     pubkey,
                     lamports,
@@ -776,8 +785,7 @@ impl LiteSVM {
             false,
             self.fee_structure.lamports_per_signature,
             0,
-            self.feature_set
-                .is_active(&remove_rounding_in_fee_calculation::id()),
+            solana_fee::FeeFeatures::from(self.feature_set.as_ref()),
         );
         let mut validated_fee_payer = false;
         let mut payer_key = None;
@@ -786,7 +794,7 @@ impl LiteSVM {
             .enumerate()
             .map(|(i, key)| {
                 let mut account_found = true;
-                let account = if solana_sdk::sysvar::instructions::check_id(key) {
+                let account = if solana_sdk_ids::sysvar::instructions::check_id(key) {
                     construct_instructions_account(message)
                 } else {
                     let instruction_account = u8::try_from(i)
@@ -825,7 +833,7 @@ impl LiteSVM {
 
                 Ok((*key, account))
             })
-            .collect::<solana_sdk::transaction::Result<Vec<_>>>();
+            .collect::<solana_transaction_error::TransactionResult<Vec<_>>>();
         let mut accounts = match maybe_accounts {
             Ok(accs) => accs,
             Err(e) => {
@@ -890,7 +898,7 @@ impl LiteSVM {
         match maybe_program_indices {
             Ok(program_indices) => {
                 let mut context = self.create_transaction_context(compute_budget, accounts);
-                let mut tx_result = MessageProcessor::process_message(
+                let mut tx_result = process_message(
                     tx.message(),
                     &program_indices,
                     &mut InvokeContext::new(
@@ -898,10 +906,10 @@ impl LiteSVM {
                         &mut program_cache_for_tx_batch,
                         EnvironmentConfig::new(
                             *blockhash,
-                            None,
-                            None,
-                            self.feature_set.clone(),
+                            self.fee_structure.lamports_per_signature,
                             0,
+                            &|_| 0,
+                            self.feature_set.clone(),
                             &self.accounts.sysvar_cache,
                         ),
                         Some(log_collector),
@@ -1249,7 +1257,7 @@ impl LiteSVM {
     fn check_transaction_age_inner(
         &self,
         tx: &SanitizedTransaction,
-    ) -> solana_sdk::transaction::Result<()> {
+    ) -> solana_transaction_error::TransactionResult<()> {
         let recent_blockhash = tx.message().recent_blockhash();
         if recent_blockhash == &self.latest_blockhash
             || self.check_transaction_for_nonce(
@@ -1273,7 +1281,10 @@ impl LiteSVM {
             .get_durable_nonce()
             .and_then(|nonce_address| self.accounts.get_account(nonce_address))
             .and_then(|nonce_account| {
-                nonce_account::verify_nonce_account(&nonce_account, message.recent_blockhash())
+                solana_nonce_account::verify_nonce_account(
+                    &nonce_account,
+                    message.recent_blockhash(),
+                )
             })
             .is_some_and(|nonce_data| {
                 message
@@ -1328,7 +1339,7 @@ fn execute_tx_helper(
     ctx: TransactionContext,
 ) -> (
     Signature,
-    solana_sdk::transaction_context::TransactionReturnData,
+    solana_transaction_context::TransactionReturnData,
     InnerInstructionsList,
     Vec<(Pubkey, AccountSharedData)>,
 ) {
@@ -1375,7 +1386,7 @@ fn validate_fee_payer(
     payer_index: IndexOfAccount,
     rent: &Rent,
     fee: u64,
-) -> solana_sdk::transaction::Result<()> {
+) -> solana_transaction_error::TransactionResult<()> {
     if payer_account.lamports() == 0 {
         error!("Payer account {payer_address} not found.");
         return Err(TransactionError::AccountNotFound);
@@ -1389,7 +1400,7 @@ fn validate_fee_payer(
         SystemAccountKind::Nonce => {
             // Should we ever allow a fees charge to zero a nonce account's
             // balance. The state MUST be set to uninitialized in that case
-            rent.minimum_balance(solana_sdk::nonce::State::size())
+            rent.minimum_balance(solana_nonce::state::State::size())
         }
     };
 
@@ -1425,8 +1436,8 @@ fn check_rent_state_with_account(
     post_rent_state: &RentState,
     address: &Pubkey,
     account_index: IndexOfAccount,
-) -> solana_sdk::transaction::Result<()> {
-    if !solana_sdk::incinerator::check_id(address)
+) -> solana_transaction_error::TransactionResult<()> {
+    if !solana_sdk_ids::incinerator::check_id(address)
         && !post_rent_state.transition_allowed_from(pre_rent_state)
     {
         let account_index = account_index as u8;