+++ crates/litesvm/src/utils/mod.rs
@@ -1,8 +1,11 @@
-use solana_sdk::{
-    account::{Account, AccountSharedData},
-    hash::{Hash, Hasher},
-    message::SanitizedMessage,
-    sysvar::{self, instructions::construct_instructions_data},
+use {
+    solana_account::{
+        Account, AccountSharedData, InheritableAccountFields, DUMMY_INHERITABLE_ACCOUNT_FIELDS,
+    },
+    solana_hash::Hash,
+    solana_instructions_sysvar::construct_instructions_data,
+    solana_message::SanitizedMessage,
+    solana_sha256_hasher::Hasher,
 };
 
 pub mod inner_instructions;
@@ -20,7 +23,24 @@ pub fn create_blockhash(bytes: &[u8]) -> Hash {
 pub fn construct_instructions_account(message: &SanitizedMessage) -> AccountSharedData {
     AccountSharedData::from(Account {
         data: construct_instructions_data(&message.decompile_instructions()),
-        owner: sysvar::id(),
+        owner: solana_sdk_ids::sysvar::id(),
         ..Account::default()
     })
+}
+
+pub(crate) fn create_loadable_account_with_fields(
+    name: &str,
+    (lamports, rent_epoch): InheritableAccountFields,
+) -> AccountSharedData {
+    AccountSharedData::from(Account {
+        lamports,
+        owner: solana_sdk_ids::native_loader::id(),
+        data: name.as_bytes().to_vec(),
+        executable: true,
+        rent_epoch,
+    })
+}
+
+pub(crate) fn create_loadable_account_for_test(name: &str) -> AccountSharedData {
+    create_loadable_account_with_fields(name, DUMMY_INHERITABLE_ACCOUNT_FIELDS)
 }