+++ crates/node-litesvm/src/transaction_error.rs
@@ -2,10 +2,8 @@ use {
     crate::to_string_js,
     core::fmt,
     napi::bindgen_prelude::{Either3, Either5},
-    solana_sdk::{
-        instruction::InstructionError as InstructionErrorOriginal,
-        transaction::TransactionError as TransactionErrorOriginal,
-    },
+    solana_instruction::error::InstructionError as InstructionErrorOriginal,
+    solana_transaction_error::TransactionError as TransactionErrorOriginal,
 };
 
 #[derive(Clone, Debug)]
@@ -298,6 +296,7 @@ pub enum TransactionErrorFieldless {
     InvalidLoadedAccountsDataSizeLimit,
     UnbalancedTransaction,
     ProgramCacheHitMaxLimit,
+    CommitCancelled,
 }
 
 to_string_js!(TransactionErrorFieldless);
@@ -489,6 +488,9 @@ pub(crate) fn convert_transaction_error(w: TransactionErrorOriginal) -> Transact
         }
         TransactionErrorOriginal::ProgramCacheHitMaxLimit => {
             TransactionError::A(TransactionErrorFieldless::ProgramCacheHitMaxLimit)
+        }
+        TransactionErrorOriginal::CommitCancelled => {
+            TransactionError::A(TransactionErrorFieldless::CommitCancelled)
         }
     }
 }