+++ crates/node-litesvm/src/util.rs
@@ -1,8 +1,4 @@
-use {
-    napi::bindgen_prelude::*,
-    solana_sdk::{hash::Hash, pubkey::Pubkey},
-    std::str::FromStr,
-};
+use {napi::bindgen_prelude::*, solana_hash::Hash, solana_pubkey::Pubkey, std::str::FromStr};
 
 pub(crate) fn convert_pubkey(address: Uint8Array) -> Pubkey {
     Pubkey::try_from(address.as_ref()).unwrap()